Revert "Fix make lint"
This reverts commit 912f21ab03ab3534c917348a902b525e8f7cbce1.

+++ lib/c-api/src/wasm_c_api/wasi/mod.rs
@@ -342,7 +342,7 @@ pub unsafe extern "C" fn wasi_env_read_stdout(
     buffer: *mut c_char,
     buffer_len: usize,
 ) -> isize {
-    let inner_buffer = slice::from_raw_parts_mut(buffer as *mut _, buffer_len);
+    let inner_buffer = slice::from_raw_parts_mut(buffer as *mut _, buffer_len as usize);
     let mut store_mut = env.store.store_mut();
     let state = env.inner.data_mut(&mut store_mut).state();
 
@@ -365,7 +365,7 @@ pub unsafe extern "C" fn wasi_env_read_stderr(
     buffer: *mut c_char,
     buffer_len: usize,
 ) -> isize {
-    let inner_buffer = slice::from_raw_parts_mut(buffer as *mut _, buffer_len);
+    let inner_buffer = slice::from_raw_parts_mut(buffer as *mut _, buffer_len as usize);
     let mut store_mut = env.store.store_mut();
     let state = env.inner.data_mut(&mut store_mut).state();
     if let Ok(mut stderr) = state.stderr() {
